Add display settings for metadata view
The project now requires django-dbsettings, but now there's a nice
addition to the admin site so that you can choose which ISMI
attributes and relations you want to display in the metadata view
and how. Moreover, the attributes and relations available are
dynamically generated from the definitions in the ISMI database,
so whenever a migration is performed, if the ISMI definitions have
changed, the classes here will be updated.

It still remains to actually implement the effects of these settings.